Floating-point values are now stored as long doubles, instead of doubles. This will give us better precision when doing computations. Some routines must convert back to double, as Qt unfortunately doesn't support long double. Much to my dismay, this didn't help with the unit tests that are failing due to precision problems, even though those tests were the reason why I did this to start with :( But - long double support is going to be useful nonetheless, so there.
Also implemented changes to increase compatibility with the
OpenDocument Formula specification:

- information function unittests properly handle cell references now
- fixed wrong unitcases for functions COLUMN, COUNTA, ISNONTEXT,
ISNA
- disabled some testcases attempting to test functions/features
which are not implemented yet
- functions AREAS and ISREF now properly handle array refs
- COUNT, SUM, SUMSQ, DEVSQ and PRODUCT will now ignore cells
containing error values
- function ISREF now properly passes error values through
- function COUNTIF now requires the first parameter to be a
reference


svn path=/trunk/koffice/; revision=879219

+++ filters/kspread/opencalc/opencalcexport.cc
@@ -645,7 +645,7 @@ void OpenCalcExport::exportCells( QDomDocument & doc, QDomElement & rowElem,
       else
         cellElem.setAttribute( "table:value-type", "float" );
 
-      cellElem.setAttribute( "table:value", QString::number( numToDouble(value.asFloat()) ) );
+      cellElem.setAttribute( "table:value", QString::number( (double)numToDouble(value.asFloat()) ) );
     }
     else
     {
